Properly store -zf-size in breakpoint function
In cases where breakpoint was called within a loop of -zf-each-breakpoint, the code changes  at 6.4.0 would reset the variable to -zf-zero-breakpoint. This caused an issue in flex-grid for anything after line 218 in that loop breaking size restricted unstack, collapse, and uncollapse.

Updated code now correctly stores the -zf-size variable in @breakpoint function if it was previously set.